refactor: UserProductApiController 클래스 @RequestBody 수정
1. String 인자를 Dto 인자로 수정
2. Rest Docs 작성
