+++ src/main/java/com/gaejangmo/apiserver/model/userproduct/dto/CommentDto.java
@@ -1,0 +1,18 @@
+package com.gaejangmo.apiserver.model.userproduct.dto;
+
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Getter
+@ToString
+@NoArgsConstructor
+@EqualsAndHashCode
+public class CommentDto {
+    private String comment;
+
+    public CommentDto(final String comment) {
+        this.comment = comment;
+    }
+}
