+++ src/main/java/com/gaejangmo/apiserver/model/userproduct/dto/ProductTypeDto.java
@@ -1,0 +1,19 @@
+package com.gaejangmo.apiserver.model.userproduct.dto;
+
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Getter
+@ToString
+@NoArgsConstructor
+@EqualsAndHashCode
+public class ProductTypeDto {
+    private String productType;
+
+    public ProductTypeDto(final String productType) {
+        this.productType = productType;
+    }
+}
+
